MAINT: Moving FakeEpicsPV to its own script called pv.py.
